+++ indispensable-asn1/src/commonMain/kotlin/at/asitplus/signum/indispensable/asn1/Asn1Elements.kt
@@ -9,12 +9,38 @@ import at.asitplus.signum.indispensable.asn1.encoding.*
 import kotlinx.io.Buffer
 import kotlinx.io.Sink
 import kotlinx.io.readByteArray
+import kotlinx.serialization.KSerializer
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.descriptors.PrimitiveKind
+import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
+import kotlinx.serialization.descriptors.SerialDescriptor
+import kotlinx.serialization.encoding.Decoder
+import kotlinx.serialization.encoding.Encoder
 import kotlin.experimental.ExperimentalObjCName
 import kotlin.native.ObjCName
+
+internal object Asn1ElementHexStringSerializer: KSerializer<Asn1Element> {
+    override val descriptor: SerialDescriptor
+        get() = PrimitiveSerialDescriptor("Asn1ElementHexpoded", PrimitiveKind.STRING)
+
+    override fun serialize(
+        encoder: Encoder,
+        value: Asn1Element
+    ) {
+        encoder.encodeString(value.toDerHexString())
+    }
+
+    override fun deserialize(decoder: Decoder): Asn1Element {
+        return Asn1Element.parseFromDerHexString(decoder.decodeString())
+    }
+
+}
 
 /**
  * Base ASN.1 data class. Can either be a primitive (holding a value), or a structure (holding other ASN.1 elements)
  */
+@Suppress("SERIALIZER_TYPE_INCOMPATIBLE")
+@Serializable(with = Asn1ElementHexStringSerializer::class)
 sealed class Asn1Element(
     val tag: Tag
 ) {
@@ -494,6 +520,8 @@ object Asn1Null : Asn1Primitive(Asn1Element.Tag.NULL, byteArrayOf())
 /**
  * ASN.1 structure. Contains no data itself, but holds zero or more [children]
  */
+@Suppress("SERIALIZER_TYPE_INCOMPATIBLE")
+@Serializable(with = Asn1ElementHexStringSerializer::class)
 sealed class Asn1Structure(
     /**
      * The tag identifying this structure
@@ -608,6 +636,8 @@ sealed class Asn1Structure(
 /**
  * Explicit ASN.1 Tag. Can contain any number of [children]
  */
+@Suppress("SERIALIZER_TYPE_INCOMPATIBLE")
+@Serializable(with = Asn1ElementHexStringSerializer::class)
 class Asn1ExplicitlyTagged
 /**
  * @param tag the ASN.1 Tag to be used will be properly encoded to have [BERTags.CONSTRUCTED] and
@@ -674,6 +704,8 @@ class Asn1Sequence internal constructor(children: List<Asn1Element>) :
 /**
  * ASN1 structure (i.e. containing child nodes) with custom tag
  */
+@Suppress("SERIALIZER_TYPE_INCOMPATIBLE")
+@Serializable(with = Asn1ElementHexStringSerializer::class)
 class Asn1CustomStructure private constructor(
     tag: Tag, children: List<Asn1Element>, sortChildren: Boolean, shouldBeSorted: Boolean
 ) :
@@ -751,6 +783,7 @@ class Asn1CustomStructure private constructor(
  * @param children the elements to put into this sequence
  */
 @Suppress("SERIALIZER_TYPE_INCOMPATIBLE")
+@Serializable(with = Asn1ElementHexStringSerializer::class)
 class Asn1EncapsulatingOctetString(children: List<Asn1Element>) :
     Asn1Structure(Tag.OCTET_STRING, children, sortChildren = false, shouldBeSorted = false),
     Asn1OctetString {
@@ -774,6 +807,8 @@ class Asn1EncapsulatingOctetString(children: List<Asn1Element>) :
  * When parsing, you should NOT cast to this class.
  * Cast to [Asn1OctetString] instead.
  */
+@Suppress("SERIALIZER_TYPE_INCOMPATIBLE")
+@Serializable(with = Asn1ElementHexStringSerializer::class)
 class Asn1PrimitiveOctetString(content: ByteArray) : Asn1Primitive(Tag.OCTET_STRING, content),
     Asn1OctetString {
 
@@ -791,6 +826,8 @@ class Asn1PrimitiveOctetString(content: ByteArray) : Asn1Primitive(Tag.OCTET_STR
 /**
  * ASN.1 SET 0x31 ([BERTags.SET] OR [BERTags.CONSTRUCTED])
  */
+@Suppress("SERIALIZER_TYPE_INCOMPATIBLE")
+@Serializable(with = Asn1ElementHexStringSerializer::class)
 open class Asn1Set private constructor(children: List<Asn1Element>, dontSort: Boolean) :
     Asn1Structure(Tag.SET, children, !dontSort, shouldBeSorted = true) {
 
@@ -820,6 +857,8 @@ open class Asn1Set private constructor(children: List<Asn1Element>, dontSort: Bo
  * @param children the elements to put into this set. will be automatically checked to have the same tag and sorted by value
  * @throws Asn1Exception if children are using different tags
  */
+@Suppress("SERIALIZER_TYPE_INCOMPATIBLE")
+@Serializable(with = Asn1ElementHexStringSerializer::class)
 class Asn1SetOf @Throws(Asn1Exception::class) internal constructor(children: List<Asn1Element>) :
     Asn1Set(children.also { it ->
         if (it.any { elem -> elem.tag != it.first().tag }) throw Asn1Exception("SET OF must only contain elements of the same tag")
@@ -828,6 +867,8 @@ class Asn1SetOf @Throws(Asn1Exception::class) internal constructor(children: Lis
 /**
  * ASN.1 primitive. Holds no children, but [content] under [tag]
  */
+@Suppress("SERIALIZER_TYPE_INCOMPATIBLE")
+@Serializable(with = Asn1ElementHexStringSerializer::class)
 open class Asn1Primitive(
     tag: Tag,
     /**
@@ -906,6 +947,8 @@ open class Asn1Primitive(
  * If you are expecting arbitrary bytes, [Asn1OctetString] is the correct type to look for.
  * (Your arbitrary bytes might inadvertently be valid ASN.1!)
  */
+@Suppress("SERIALIZER_TYPE_INCOMPATIBLE")
+@Serializable(with = Asn1ElementHexStringSerializer::class)
 sealed interface Asn1OctetString {
 
     /**