+++ CHANGELOG.md
@@ -22,6 +22,7 @@
     * Add `Asn1Null` object
     * Add human-readable ASN.1 element toString
     * Make Asn1OctetString interface sealed
+* Make Asn1Element serializable
 * Strippable `KnownOIDs`
     * Move `KnownOIDs` into a discrete module `indispensable-oids`
 * OID descriptions: